set_electron bugfix and edge case. More tests
Bugfix: When entering the electron density as a list of integers,
the resulting array type used to be int64 rather than float64. This
could cause some problems. Now the dtype is forced to be float.
Edge case: When allow_fractions=False, it is possible to get
non-matching electron numbers and electron densities. The electron
density now gets scaled to match the electron number.