KagomeHubbardKPoints had the wrong complex conjugation.
I had mixed up the ordering of indices going from tensor to matrix
operations. The only tangible effect this would have had is in changing
the complex conjugation of the eigenvectors, so it won't affect any of
my results so far, but it's worth catching now before it is too late.

tests have been updated to reflect this change.