Mixing routines now return the residual. utils.converge optimised
linear_mixing still defaults to returning the energy, but there is a
flag to change its behaviour.
Anderson mixing, even though I don't use it, has been updated to accept
GCE mixing, in line with linear and Pulay mixing.

utils.converge takes advantage of receivintg the residual such that it
does not have to recompute the residual while checking for convergence
or printing progress. This is an appreciable saving for large systems
and small values of interval.